requirements.txt
fastapi
uvicorn
python-dotenv
chromadb
sentence-transformers
google-generativeai
PyPDF2
python-docx